Remove add/delete lists from weapon (important)
Removes the add/delete lists. Weapon is now Updateable and can
access the container from its update method. However, the fire() and
remove() methods do not have access per se. I added the container
manually to get it done with, but this should be updated in a way
that these methods are called from the update method.