Add ResourceWrapper to AbstractGameObject
Changes the constructor to no longer require an image (this image
would have to be generated from the subclass with the resourcewrapper
in the AbstractGameObject superclass whose constructor is being called,
but that is not possible).

Instead, the constructor uses the abstract method setDefaultImage to
determine the desired image. The subclass is responsible for
correct implementation of its image.

Also adds the setter for the resourceWrapper. This allows a
mocked wrapper to be set instead of the default.

< interface, organisation