Bugfix not working menu buttons
The left mouse button press is removed from the input buffer after
calling isMousePressed(). This means that only the first button in the
update cycle will work (highscores in this caes).

Using isMouseButtonDown() instead polls the button state. If it is down
at that moment, regardless of any previous states, it will return true.

< bug