Replace MouseButtonDown with MousePressed
This requires one global check on mousePressed, as it clears the
buffer of that button press and would therefore only fire for the
first button to be checked (the buffer would already be empty
by the time the other updates were reached).

Rewrote tests to reflect new mocks, removes unnecessary tests
(isAreaClicked() has been removed from the controller), adds tests
to reach newly created branches (if statement in update for buttonpress).

< testing, organistion, bug, interface